Découpage du context Behat
La classe de context commence à devenir un peu difficile à gérer avec
les ajouts récents de phrases. La recommandation officielle est de
découper la classe avec des traits.
